requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.1
flask-cors==6.0.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
smmap==5.0.2
urllib3==2.4.0
URLObject==2.4.3
Werkzeug==3.1.3
zipp==3.22.0
